Revert base css to original, move changes to local.css, and fix topbar and content width.
This correctly fixes the topbar and content width doing what commit 54029 did in the correct way.  That change forced the topbar width to be too small when additional categories were adding, making it take up two lines and overlapping the top of the content and sidebar areas.  This also moves changes to local.css where they should probably be placed.